move navigation components into own folder
