+++ stockpile-web/src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import { Route, Switch, BrowserRouter as Router } from 'react-router-dom';
-import Navigation from './components/common/Navigation';
+import Navigation from './components/pages/common/Navigation';
 import Home from './components/pages/Home';
 import ArticleListing from './components/pages/article/ArticleListing';
 import DataImport from './components/pages/DataImport';
