+++ stockpile-web/src/components/common/navigation/NavBarItem.js
