+++ stockpile-web/src/components/pages/common/Navigation.js
@@ -1,5 +1,5 @@
-import NavBar from './NavBar'
-import NavBarItem from './NavBarItem'
+import NavBar from '../../common/navigation/NavBar'
+import NavBarItem from '../../common/navigation/NavBarItem'
 
 const Navigation = ({brand}) => {
     return (
@@ -7,8 +7,8 @@ const Navigation = ({brand}) => {
             <NavBarItem id="navHome" label="Home" link="/" active="true"></NavBarItem>
             <NavBarItem id="navArticle" label="Article" link="/articles"></NavBarItem>
             <NavBarItem id="navUpload" label="Upload" link="/data-import"></NavBarItem>
-            <NavBarItem id="navAdmin" label="Admin" link="#">
-                <NavBarItem id="navConfigureUploadSchema" label="Configure upload schema" link="/admin/configure-upload-schemas" />
+            <NavBarItem id="navAdmin" label="Administration" link="#">
+                <NavBarItem id="navConfigureUploadSchema" label="Upload schema" link="/admin/configure-upload-schemas" />
             </NavBarItem>
             <NavBarItem id="navAboutUs" label="About" link="/about" />
         </NavBar>
